Remoção completa do s!settings

+++ src/main/java/com/kuuhaku/command/commands/discord/moderation/MakeLogCommand.java
@@ -23,11 +23,10 @@ import com.kuuhaku.command.Executable;
 import com.kuuhaku.controller.postgresql.GuildDAO;
 import com.kuuhaku.model.annotations.Command;
 import com.kuuhaku.model.annotations.Requires;
-import com.kuuhaku.model.persistent.GuildConfig;
+import com.kuuhaku.model.persistent.guild.GuildConfig;
 import com.kuuhaku.utils.Helper;
 import net.dv8tion.jda.api.Permission;
 import net.dv8tion.jda.api.entities.*;
-import net.dv8tion.jda.api.exceptions.InsufficientPermissionException;
 
 @Command(
 		name = "criarlog",
@@ -39,25 +38,20 @@ public class MakeLogCommand implements Executable {
 
 	@Override
 	public void execute(User author, Member member, String command, String argsAsText, String[] args, Message message, TextChannel channel, Guild guild, String prefix) {
+		GuildConfig gc = GuildDAO.getGuildById(guild.getId());
+
 		try {
-
-			GuildConfig gc = com.kuuhaku.controller.postgresql.GuildDAO.getGuildById(guild.getId());
-			try {
-				TextChannel tc = guild.getTextChannelById(gc.getCanalLog());
-				if (tc != null)
-					tc.delete().complete();
-			} catch (Exception ignore) {
-			}
-
-			guild.createTextChannel("shiro-log")
-					.setParent(channel.getParent())
-					.queue(c -> {
-						gc.setCanalLog(c.getId());
-						channel.sendMessage("✅ | Canal de log criado com sucesso em " + c.getAsMention()).queue(null, Helper::doNothing);
-						GuildDAO.updateGuildSettings(gc);
-					});
-		} catch (InsufficientPermissionException e) {
-			channel.sendMessage("❌ | Não tenho permissões sufficientes para criar um canal.").queue();
+			TextChannel tc = gc.getLogChannel();
+			if (tc != null) tc.delete().complete();
+		} catch (Exception ignore) {
 		}
+
+		guild.createTextChannel("shiro-log")
+				.setParent(channel.getParent())
+				.queue(c -> {
+					gc.setLogChannel(c.getId());
+					channel.sendMessage("✅ | Canal de log criado com sucesso em " + c.getAsMention()).queue(null, Helper::doNothing);
+					GuildDAO.updateGuildSettings(gc);
+				});
 	}
 }
